+++ test/cli.test.js
@@ -1,36 +1,28 @@
-const path = require('path');
-const test = require('tape');
-const sinon = require('sinon');
-const cli = require('../lib/cli');
-const cfnConfig = require('..');
+import path from 'path';
+import test from 'tape';
+import sinon from 'sinon';
+import {
+    parse,
+    main
+} from '../lib/cli.js';
+import cfnConfig from '../index.js';
 
 test('[cli.parse] aliases and defaults', (t) => {
     const args = ['create', 'testing', 'relative/path', '-c', 'config', '-t', 'template'];
-    const parsed = cli.parse(args, {});
+    const parsed = parse(args, {});
 
     t.equal(parsed.command, 'create', 'contains command');
     t.equal(parsed.environment, 'testing', 'contains environment');
     t.equal(parsed.templatePath, path.resolve('relative/path'), 'contains absolute template path');
     t.deepEqual(parsed.options, {
-        d: false,
         decrypt: false,
-        e: false,
         extended: false,
-        f: false,
         force: false,
-        k: false,
         kms: false,
-        n: path.basename(process.cwd()),
         name: path.basename(process.cwd()),
-        r: 'us-east-1',
         region: 'us-east-1',
-        t: 'template',
         templateBucket: 'template',
-        c: 'config',
         configBucket: 'config',
-        p: undefined,
-        parameters: undefined,
-        x: false,
         expand: false
     }, 'provided expected options');
     t.deepEqual(parsed.overrides, { force: false, kms: false, parameters: undefined, expand: false }, 'provided expected overrides');
@@ -50,28 +42,19 @@ test('[cli.parse] sets options', (t) => {
         '-p', '{}'
     ];
 
-    const parsed = cli.parse(args, {});
+    const parsed = parse(args, {});
 
     t.deepEqual(parsed.options, {
-        d: true,
         decrypt: true,
-        e: true,
         extended: true,
-        f: true,
         force: true,
-        k: 'kms-id',
         kms: 'kms-id',
-        n: 'my-stack',
         name: 'my-stack',
-        r: 'eu-west-1',
         region: 'eu-west-1',
-        t: 'template',
         templateBucket: 'template',
-        c: 'config',
         configBucket: 'config',
-        p: {},
         parameters: {},
-        x: true,
+        p: {},
         expand: true
     }, 'provided expected options');
     t.deepEqual(parsed.overrides, { force: true, kms: 'kms-id', parameters: {}, expand: true }, 'provided expected overrides');
@@ -80,23 +63,23 @@ test('[cli.parse] sets options', (t) => {
 });
 
 test('[cli.parse] handles default template bucket on create & update', (t) => {
-    let parsed = cli.parse(['info', 'testing'], {});
+    let parsed = parse(['info', 'testing'], {});
     t.notOk(parsed.options.templateBucket, 'not set when not needed');
 
-    parsed = cli.parse(['create', 'testing'], { AWS_ACCOUNT_ID: '123456789012' });
+    parsed = parse(['create', 'testing'], { AWS_ACCOUNT_ID: '123456789012' });
     t.equal(parsed.options.templateBucket, 'cfn-config-templates-123456789012-us-east-1', 'uses default for create');
 
-    parsed = cli.parse(['update', 'testing'], { AWS_ACCOUNT_ID: '123456789012' });
+    parsed = parse(['update', 'testing'], { AWS_ACCOUNT_ID: '123456789012' });
     t.equal(parsed.options.templateBucket, 'cfn-config-templates-123456789012-us-east-1', 'uses default for create');
 
     t.throws(
-        function() { cli.parse(['create', 'testing'], {}); },
+        function() { parse(['create', 'testing'], {}); },
         /Provide \$AWS_ACCOUNT_ID as an environment variable to use the default template bucket, or set --template-bucket/,
         'throws error on create without $AWS_ACCOUNT_ID'
     );
 
     t.throws(
-        function() { cli.parse(['update', 'testing'], {}); },
+        function() { parse(['update', 'testing'], {}); },
         /Provide \$AWS_ACCOUNT_ID as an environment variable to use the default template bucket, or set --template-bucket/,
         'throws error on update without $AWS_ACCOUNT_ID'
     );
@@ -125,7 +108,7 @@ test('[cli.main] no command', async(t) => {
     const parsed = Object.assign({}, base, { command: undefined });
 
     try {
-        await cli.main(parsed);
+        await main(parsed);
         t.fail();
     } catch (err) {
         t.equal(err.message, 'Error: invalid command\n\n' + parsed.help, 'expected error message');
@@ -137,7 +120,7 @@ test('[cli.main] bad command', async(t) => {
     const parsed = Object.assign({}, base, { command: 'hibbity' });
 
     try {
-        await cli.main(parsed);
+        await main(parsed);
         t.fail();
     } catch (err) {
         t.equal(err.message, 'Error: invalid command\n\n' + parsed.help, 'expected error message');
@@ -150,7 +133,7 @@ test('[cli.main] no environment', async(t) => {
     const parsed = Object.assign({}, base, { environment: undefined });
 
     try {
-        await cli.main(parsed);
+        await main(parsed);
         t.fail();
     } catch (err) {
         t.equal(err.message, 'Error: missing environment\n\n' + parsed.help, 'expected error message');
@@ -163,7 +146,7 @@ test('[cli.main] no template path (create)', async(t) => {
     const parsed = Object.assign({}, base, { templatePath: undefined });
 
     try {
-        await cli.main(parsed);
+        await main(parsed);
         t.fail();
     } catch (err) {
         t.equal(err.message, 'Error: missing templatePath\n\n' + parsed.help, 'expected error message');
@@ -184,7 +167,7 @@ test('[cli.main] no template path (info)', async(t) => {
     const parsed = Object.assign({}, base, { command: 'info', templatePath: undefined });
 
     try {
-        await cli.main(parsed);
+        await main(parsed);
     } catch (err) {
         t.error(err);
     }
@@ -208,7 +191,7 @@ test('[cli.main] create', async(t) => {
     });
 
     try {
-        await cli.main(base);
+        await main(base);
     } catch (err) {
         t.error(err);
     }
@@ -233,7 +216,7 @@ test('[cli.main] update', async(t) => {
     const parsed = Object.assign({}, base, { command: 'update' });
 
     try {
-        await cli.main(parsed);
+        await main(parsed);
     } catch (err) {
         t.error(err);
     }
@@ -257,7 +240,7 @@ test('[cli.main] delete', async(t) => {
     const parsed = Object.assign({}, base, { command: 'delete' });
 
     try {
-        await cli.main(parsed);
+        await main(parsed);
     } catch (err) {
         t.error(err);
     }
@@ -283,7 +266,7 @@ test('[cli.main] info', async(t) => {
     const parsed = Object.assign({}, base, { command: 'info' });
 
     try {
-        await cli.main(parsed);
+        await main(parsed);
     } catch (err) {
         t.error(err);
     }
@@ -308,7 +291,7 @@ test('[cli.main] save (with kms)', async(t) => {
     parsed.overrides.kms = true;
 
     try {
-        await cli.main(parsed);
+        await main(parsed);
     } catch (err) {
         t.error(err);
     }
@@ -332,7 +315,7 @@ test('[cli.main] save (without kms)', async(t) => {
     const parsed = Object.assign({}, base, { command: 'save' });
 
     try {
-        await cli.main(parsed);
+        await main(parsed);
     } catch (err) {
         t.error(err);
     }