+++ test/fixtures/huge-template.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
     Resources: new Array(5000).fill(0).reduce((resources, _, i) => {
         resources['Thing' + i] = {
             Type: 'AWS::SNS::Topic',